career-pulse-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-cloud-storage>=2.10.0
google-api-python-client>=2.125.0
google-auth>=2.35.0
pandas>=2.0.0
requests>=2.31.0
